+++ vllm/distributed/kv_transfer/kv_connector/v1/nixl/__init__.py
@@ -2,6 +2,15 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 """NIXL KV-cache transfer connector (disaggregated prefill / decode)."""
 
+from vllm.distributed.kv_transfer.kv_connector.v1.nixl.base_connector import (
+    NixlBaseConnector,
+)
+from vllm.distributed.kv_transfer.kv_connector.v1.nixl.base_scheduler import (
+    NixlBaseConnectorScheduler,
+)
+from vllm.distributed.kv_transfer.kv_connector.v1.nixl.base_worker import (
+    NixlBaseConnectorWorker,
+)
 from vllm.distributed.kv_transfer.kv_connector.v1.nixl.connector import (
     NixlConnector,
 )
@@ -9,6 +18,15 @@ from vllm.distributed.kv_transfer.kv_connector.v1.nixl.metadata import (
     NixlAgentMetadata,
     NixlConnectorMetadata,
     NixlHandshakePayload,
+)
+from vllm.distributed.kv_transfer.kv_connector.v1.nixl.pull_connector import (
+    NixlPullConnector,
+)
+from vllm.distributed.kv_transfer.kv_connector.v1.nixl.pull_scheduler import (
+    NixlPullConnectorScheduler,
+)
+from vllm.distributed.kv_transfer.kv_connector.v1.nixl.pull_worker import (
+    NixlPullConnectorWorker,
 )
 from vllm.distributed.kv_transfer.kv_connector.v1.nixl.scheduler import (
     NixlConnectorScheduler,
@@ -22,10 +40,16 @@ from vllm.distributed.kv_transfer.kv_connector.v1.nixl.worker import (
 
 __all__ = [
     "NixlAgentMetadata",
+    "NixlBaseConnector",
+    "NixlBaseConnectorScheduler",
+    "NixlBaseConnectorWorker",
     "NixlConnector",
     "NixlConnectorMetadata",
     "NixlConnectorScheduler",
     "NixlConnectorWorker",
     "NixlHandshakePayload",
     "NixlKVConnectorStats",
+    "NixlPullConnector",
+    "NixlPullConnectorScheduler",
+    "NixlPullConnectorWorker",
 ]
